crop-inventory-service/requirements.txt
fastapi
uvicorn
python-dotenv
py_eureka_client
sqlalchemy
pymysql